pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "txt2detection"

version = "0.0.1"
authors = [ 
  { name="DOGESEC", email="support@dogesec.com" },
]
description = "txt2detection is a tool"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "stix2",
    "python-arango>=8.1.3; python_version >= '3.8'",
    "tqdm>=4.66.4; python_version >= '3.7'",
    "jsonschema>=4.22.0; python_version >= '3.8'",
    "requests>=2.31.0; python_version >= '3.7'",
    "python-dotenv>=1.0.1",
    "pyyaml",
  ]
[project.urls]
Homepage = "https://github.com/muchdogesec/txt2detection"
Issues = "https://github.com/muchdogesec/txt2detection/issues"

[project.scripts]
txt2detection = "txt2detection.__main__:main"

[project.optional-dependencies]
llms = [
  "llama-index-core==0.12.7",
  "llama-index-llms-anthropic==0.6.3",
  "llama-index-llms-gemini==0.4.2",
  "llama-index-llms-openai==0.3.11",
  "llama-index-llms-openai-like==0.3.3",
  "llama-index-llms-deepseek==0.1.1",
  "llama-index-llms-openrouter==0.3.1",
]


[tool.hatch.build.targets.wheel.force-include]
"config" = "txt2detection/config"